2025-day-3 part two successfully completed

+++ advent-of-code-2025/day3/src/Lobby.java
@@ -204,7 +204,7 @@ public class Lobby {
                 """;
 
         String[] banks = input.split("\n");
-        int totalJoltage = 0;
+        long totalJoltage = 0;
         for (String bank : banks) {
             if (bank.trim().isEmpty()) {
                 continue;
